+++ src/main/resources/META-INF/services/javax.script.ScriptEngineFactory
@@ -1,1 +1,1 @@
-org.hosh.script.HoshScriptEngineFactory
+hosh.script.HoshScriptEngineFactory
